refactor controller and frame to have shared context
